+++ src/deigde/editors/project/testrun/projTestRunProcess.cpp
@@ -19,17 +19,19 @@
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
+#include <dragengine/dragengine_configuration.h>
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <unistd.h>
 #include <signal.h>
 #include <stdint.h>
-#include <sys/time.h>
 #include <sys/types.h>
 #ifdef OS_W32
 #include <dragengine/app/include_windows.h>
 #else
+#include <unistd.h>
+#include <sys/time.h>
 #include <sys/wait.h>
 #include <sys/select.h>
 #endif
@@ -132,7 +134,7 @@ void projTestRunProcess::WriteFloatToPipe( float value ){
 }
 
 void projTestRunProcess::WriteString16ToPipe( const char *string ){
-	const int length = strlen( string );
+	const int length = ( int )strlen( string );
 	WriteUShortToPipe( length );
 	WriteToPipe( string, length );
 }
@@ -171,7 +173,7 @@ int projTestRunProcess::ReadUShortFromPipe(){
 	return vushort;
 }
 
-int projTestRunProcess::ReadFloatFromPipe(){
+float projTestRunProcess::ReadFloatFromPipe(){
 	float value;
 	ReadFromPipe( &value, sizeof( float ) );
 	return value;
